simplifying package structure, pip installalled package now works
