+++ examples/learn_kinematics_of_iiwa.py
@@ -9,6 +9,7 @@ from hydra.experimental import initialize_config_dir
 
 from differentiable_robot_model.differentiable_robot_model import DifferentiableRobotModel, DifferentiableKUKAiiwa
 from differentiable_robot_model.data_generation_utils import generate_random_forward_kinematics_data
+import diff_robot_data
 
 torch.set_printoptions(precision=3, sci_mode=False)
 random.seed(0)
@@ -16,7 +17,7 @@ np.random.seed(1)
 torch.manual_seed(0)
 
 
-abs_config_dir = os.path.abspath("../conf")
+abs_config_dir = os.path.abspath("conf")
 # we load a learnable robot model
 with initialize_config_dir(config_dir=abs_config_dir):
     # which parameters are learnable is specified in the config file
@@ -25,7 +26,10 @@ with initialize_config_dir(config_dir=abs_config_dir):
     )
 
 gt_robot_model = DifferentiableKUKAiiwa()
-learnable_robot_model = DifferentiableRobotModel(**learnable_robot_model_cfg.model)
+urdf_path = os.path.join(diff_robot_data.__path__[0], learnable_robot_model_cfg.model.rel_urdf_path)
+learnable_robot_model = DifferentiableRobotModel(urdf_path,
+                                                 learnable_robot_model_cfg.model.learnable_rigid_body_config,
+                                                 learnable_robot_model_cfg.model.name)
 
 train_data = generate_random_forward_kinematics_data(gt_robot_model, n_data=100, ee_name="iiwa_link_ee")
 q = train_data["q"]
